Add interfaces for Ciudadano, Pregunta, and Voto
Added new TypeScript interfaces for Ciudadano, Pregunta, and Voto in the src/interfaces directory. These interfaces define the structure and properties of the corresponding data objects used in the application.

Also added an index.ts file in the src/interfaces directory to export the interfaces for easier import in other files.

Closes #123